Refactor: Modernize codebase and add Swagger documentation
- I refactored the project structure by moving all source code into a `src` directory.
- I updated `package.json` to reflect the new project structure.
- I added Swagger documentation to all API routes for better API discoverability and testing.
