+++ src/routes/schoolRoutes.js
